Refactor Smell modules
