+++ spec/DuplicationSpec.hs
@@ -1,7 +1,7 @@
 module DuplicationSpec (spec) where
 
 import           Test.Hspec
-import           Language.Mulang.Inspector.Generic.Duplication
+import           Language.Mulang.Inspector.Smell.Duplication
 import           Language.Mulang.Parsers.JavaScript
 
 spec :: Spec
